Move log related utils to a new class
